+++ src/azure-cli/HISTORY.rst
@@ -22,7 +22,7 @@ Release History
 
 **Storage**
 
-*az storage account create/update: Add --enable-large-file-share parameter to support large file shares for storage account.
+* az storage account create/update: Add --enable-large-file-share parameter to support large file shares for storage account.
 
 2.0.74
 ++++++